requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
shap
matplotlib
google-generativeai